fix(store/postgres): use int16 for SMALLINT bool columns to fix pgx v5 binary encoding
pgx v5 in binary mode cannot encode Go bool or int into int2 (SMALLINT).
Wrap all boolean fields written to SMALLINT columns with boolToInt() which
now returns int16, the type pgx v5 can encode to int2 without a plan error.

Affected fields: Detached, WebPTYEnabled (agents), AutoProvide (runtime_brokers),
Secret (env_vars), Locked (templates, harness_configs).

All 21 conformance subtests now pass against a real Postgres instance.

+++ pkg/store/postgres/helpers.go
@@ -95,7 +95,7 @@ func timeFromPtr(t *time.Time) time.Time {
 	return *t
 }
 
-func boolToInt(b bool) int {
+func boolToInt(b bool) int16 {
 	if b {
 		return 1
 	}
